+++ galley/pkg/config/processing/snapshotter/analyzingdistributor_test.go
@@ -22,9 +22,6 @@ import (
 	"istio.io/istio/galley/pkg/config/analysis"
 	"istio.io/istio/galley/pkg/config/analysis/diag"
 	coll "istio.io/istio/galley/pkg/config/collection"
-	"istio.io/istio/galley/pkg/config/event"
-	"istio.io/istio/galley/pkg/config/processing"
-	"istio.io/istio/galley/pkg/config/processing/transformer"
 	"istio.io/istio/galley/pkg/config/resource"
 	"istio.io/istio/galley/pkg/config/schema/collection"
 	"istio.io/istio/galley/pkg/config/testing/data"
@@ -72,7 +69,7 @@ func TestAnalyzeAndDistributeSnapshots(t *testing.T) {
 		collectionAccessed = col
 	}
 
-	ad := NewAnalyzingDistributor(u, analysis.Combine("testCombined", a), d, cr, collection.Names{}, transformer.Providers{})
+	ad := NewAnalyzingDistributor(u, analysis.Combine("testCombined", a), d, cr)
 
 	sDefault := getTestSnapshot("a", "b")
 	sSynthetic := getTestSnapshot("c")
@@ -103,34 +100,4 @@ func getTestSnapshot(names ...string) *Snapshot {
 	return &Snapshot{
 		set: coll.NewSetFromCollections(c),
 	}
-}
-
-func TestGetDisabledOutputs(t *testing.T) {
-	g := NewGomegaWithT(t)
-
-	in1 := collection.NewName("in1")
-	in2 := collection.NewName("in2")
-	in3 := collection.NewName("in3")
-	in4 := collection.NewName("in4")
-	in5 := collection.NewName("in5")
-	out1 := collection.NewName("out1")
-	out2 := collection.NewName("out2")
-	out3 := collection.NewName("out3")
-	out4 := collection.NewName("out4")
-
-	blankFn := func(_ processing.ProcessorOptions) event.Transformer {
-		return event.NewFnTransform(collection.Names{}, collection.Names{}, func() {}, func() {}, func(e event.Event, handler event.Handler) {})
-	}
-
-	xformProviders := transformer.Providers{
-		transformer.NewProvider(collection.Names{in1}, collection.Names{out1, out2}, blankFn),
-		transformer.NewProvider(collection.Names{in2}, collection.Names{out3}, blankFn),
-		transformer.NewProvider(collection.Names{in3}, collection.Names{out3}, blankFn),
-		transformer.NewProvider(collection.Names{in4, in5}, collection.Names{out4}, blankFn),
-	}
-
-	g.Expect(getDisabledOutputs(collection.Names{in1}, xformProviders)).To(ConsistOf(out1, out2))
-	g.Expect(getDisabledOutputs(collection.Names{in2}, xformProviders)).To(ConsistOf())
-	g.Expect(getDisabledOutputs(collection.Names{in2, in3}, xformProviders)).To(ConsistOf(out3))
-	g.Expect(getDisabledOutputs(collection.Names{in4}, xformProviders)).To(ConsistOf(out4))
 }
