+++ galley/pkg/config/schema/schema.go
@@ -188,8 +188,8 @@ func (k KubeResources) MustFind(group, kind string) KubeResource {
 	return r
 }
 
-// GetDisabled returns the names of disabled collections
-func (k KubeResources) GetDisabled() collection.Names {
+// DisabledCollections returns the names of disabled collections
+func (k KubeResources) DisabledCollections() collection.Names {
 	disabledCollections := make([]collection.Name, 0)
 	for _, r := range k {
 		if r.Disabled {
